% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{add_exp_data_kallisto}
\alias{add_exp_data_kallisto}
\title{Adds experiment data from Kallisto estimated reads in the form of a SummarizedExperiment}
\usage{
add_exp_data_kallisto(
  time_object,
  sample_dta_path,
  kallisto_files,
  tx2gene_path,
  matrix_name = "raw"
)
}
\arguments{
\item{time_object}{A TimeSeries_Object}

\item{sample_dta_path}{String which gives the csv path to the sample data}

\item{kallisto_files}{a vector containing the path(s) to the kallisto estimated read
files. Each path must contain the associated sample name.}

\item{tx2gene_path}{A path to a csv containing a dataframe showing transcript
to gene conversion (likely ENST to ENSG) The format of the dataframe will depend
on the kallisto files inputed}

\item{matrix_name}{Name to be give to the matrix. Default is 'raw'}
}
\value{
time_object - the updated timeSeries object
}
\description{
A means to add experiment data from Kallisto to TiSA objects.
The function utilizes DESeq2 to convert transcripts to genes along with a
transcript to gene conversion dataframe.
DESeq2 creates a dds object, the necessary elements are extracted and stored into
a timeseries object instead.

This function, if used, is intended to replace the standard \code{add_experiment_data} function.
}
