% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{add_exp_data_tximport}
\alias{add_exp_data_tximport}
\title{Adds experiment data from tximport estimated reads in the form of a SummarizedExperiment}
\usage{
add_exp_data_tximport(
  time_object,
  sample_dta_path,
  alt_files,
  tx2gene_path,
  tx_type,
  matrix_name = "raw"
)
}
\arguments{
\item{time_object}{A TimeSeries_Object}

\item{sample_dta_path}{String which gives the csv path to the sample data}

\item{alt_files}{a vector containing the path(s) to the txfile reads
files. Each path must contain the associated sample name.}

\item{tx2gene_path}{A path to a csv containing a dataframe showing transcript
to gene conversion. The format of the dataframe will depend
on the kallisto files inputed}

\item{tx_type}{A string detailing from what tool the experiment comes from. For
example, 'salmon', or 'kallisto'}

\item{matrix_name}{Name to be give to the matrix. Default is 'raw'}
}
\value{
time_object - the updated timeSeries object
}
\description{
A means to add experiment data from Kallisto or Salmon to TiSA objects.
The function utilizes DESeq2 to convert transcripts to genes along with a
transcript to gene conversion dataframe.
DESeq2 creates a dds object, the necessary elements are extracted and stored into
a timeseries object instead.

This function, if used, is intended to replace the standard \code{add_experiment_data} function.
}
