% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{add_experiment_data}
\alias{add_experiment_data}
\title{Adds experiment data in the form of a SummarizedExperiment}
\usage{
add_experiment_data(
  time_object,
  sample_dta_path,
  count_dta_path,
  limma_ID_replace = "GeneName"
)
}
\arguments{
\item{time_object}{A TimeSeries_Object}

\item{sample_dta_path}{String which gives the csv path to the sample data}

\item{count_dta_path}{String which gives the path to the count or microarray data}

\item{limma_ID_replace}{The name of gene id's to use from the 'genes' dataframe
in the Elist}
}
\value{
count_matrix
}
\description{
The function creates a SummarizedExperiment from the provided data.
It reads the sample data and value files, filters them to contain only the necessary
groups and then creates a SummarizedExperiment to then store this into the TimeSeris_Object
}
\examples{
write_example_data_to_dir('PBMC')
my_path_data<-'data/PBMC/raw_counts_TS'
my_path_sample_dta<-'data/PBMC/sample_file.csv'

TS_object <- new('TimeSeries_Object',
                 group_names=c('IgM','LPS'),group_colors=graph_vect,DE_method='DESeq2',
                 DE_p_filter='padj',DE_p_thresh=0.05,DE_l2fc_thresh=1,
                 PART_l2fc_thresh=4,sem_sim_org='org.Hs.eg.db',Gpro_org='hsapiens')

TS_object <- add_experiment_data(TS_object,sample_dta_path=my_path_sample_dta,count_dta_path=my_path_data)

}
