% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{add_semantic_similarity_data}
\alias{add_semantic_similarity_data}
\title{Calculate and add semantic similarity to object}
\usage{
add_semantic_similarity_data(object, ont_sem_sim)
}
\arguments{
\item{object}{A timeseries object}

\item{ont_sem_sim}{An ontology for the semantic similarity. EX: 'BP','MF','CC'}
}
\description{
This function calculates the semantic similarity object for an
organism. It then stores this within the TimeSeries object and returns it
}
