% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AID_example_gpro_res}
\alias{AID_example_gpro_res}
\title{Gprofiler results for the function examples}
\format{
A list of lists
}
\source{
<https://www.github.com/Ylefol/TimeSeriesAnalysis>
}
\usage{
AID_example_gpro_res
}
\description{
A list containing the gprofiler results (11 clusters) for the AID example.
This data exists to allow for the creation of function examples in the event
that no connection to gprofiler could be made (host connection error)
}
\keyword{datasets}
