% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_GO_results_functions.R
\name{calculate_and_format_MDS}
\alias{calculate_and_format_MDS}
\title{Calculate MDS matrix}
\usage{
calculate_and_format_MDS(GO_df, semantic_data, measure = "Wang")
}
\arguments{
\item{GO_df}{The dataframe of GOs contained in the 'Gpro_org' slot of a time_object}

\item{semantic_data}{semantic similarity data as created by the godata function}

\item{measure}{the measuring method for semantic distance}
}
\value{
res.mds A dataframe containing the MDS dimensions, along with groupings and
colors which are equivalent to the cluster colors
}
\description{
Function which calculates semantic distance and MDS from a dataframe of GOs

The function uses a GOSemSim object to calculates the termdistance, followed
by a MDS calculation.
The resulting dataframe is formatted to be used in \code{plot_MDS}

This function is used to create a GO term plot for WGCNA modules
}
\examples{
TS_object<-create_example_object_for_R()
TS_object <- normalize_timeSeries_with_deseq2(time_object=TS_object)
#Perform conditional differential gene expression analysis
TS_object<-conditional_DE_wrapper(TS_object,vignette_run=TRUE)
TS_object<-temporal_DE_wrapper(TS_object,do_all_combinations=TRUE,vignette_run=TRUE)
#Extract genes for PART clustering based on defined log(2)foldChange threshold
signi_genes<-select_genes_with_l2fc(TS_object)

#Use all samples, but implement a custom order. In this case it is reversed
sample_data<-exp_sample_data(TS_object)
TS_groups<-slot(TS_object,'group_names')
samps_2<-sample_data$sample[sample_data$group==TS_groups[2]]
samps_1<-sample_data$sample[sample_data$group==TS_groups[1]]

#Create the matrix that will be used for PART clustering
TS_object<-prep_counts_for_PART(object=TS_object,target_genes=signi_genes,scale=TRUE,target_samples=c(samps_2,samps_1))
TS_object<-compute_PART(TS_object,part_bootstrap=10,part_min_clust=10,dist_param="euclidean", hclust_param="average",vignette_run=TRUE)
TS_object<-run_gprofiler_PART_clusters(TS_object,vignette_run=TRUE) #Run the gprofiler analysis
#Results saved to created directory
gpro_res<-gprofiler_cluster_analysis(TS_object,'GO:BP',save_path=NULL)
GO_clusters<-gpro_res[['GO_df']]
sem_dta<-slot(TS_object,'sem_list')
plot_data<-calculate_and_format_MDS(GO_clusters,sem_dta)

}
