% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{calculate_EB}
\alias{calculate_EB}
\title{Empirical bayesan analysis}
\usage{
calculate_EB(micro_arr_dta, matrix_model, comparison)
}
\arguments{
\item{micro_arr_dta}{An Elist object containing the normalized and corrected microarray data}

\item{matrix_model}{the matrix model as created by the \code{prep_matrix()}}

\item{comparison}{The name of the comparison being run}
}
\value{
The empirical Bayesian results
}
\description{
Function which uses a model matrix to calculate the empirical Bayesian results
from a provided microarray normalized and corrected Elist.
}
