% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{calculate_gene_traj_data}
\alias{calculate_gene_traj_data}
\title{Calculate trajectory of a single gene}
\usage{
calculate_gene_traj_data(time_object, target_gene, log_timepoint = FALSE)
}
\arguments{
\item{time_object}{A timeseries object}

\item{target_gene}{String indicating the gene}

\item{log_timepoint}{Boolean indicating if timepoints should be log transformed}
}
\value{
dataframe with the trajectory data for the requested gene
}
\description{
Function takes in a target gene and will calculate the genes trajectory
over the time points within the time object
}
\examples{
TS_object<-create_example_object_for_R()
TS_object <- normalize_timeSeries_with_deseq2(time_object=TS_object)
#Perform conditional differential gene expression analysis
TS_object<-conditional_DE_wrapper(TS_object)
aicda_traj_dta<-calculate_gene_traj_data(TS_object,'AICDA')

}
