% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celegans_gpro_res}
\alias{celegans_gpro_res}
\title{Gprofiler results for the celegans example dataset}
\format{
A list of lists
}
\source{
<https://www.github.com/Ylefol/TimeSeriesAnalysis>
}
\usage{
celegans_gpro_res
}
\description{
A list containing the gprofiler results (14 clusters) for the celegans example.
This data exists to allow for the creation of Vignettes in the event of a network failure.
}
\keyword{datasets}
