% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{conditional_DE_wrapper}
\alias{conditional_DE_wrapper}
\title{Create conditional DE results}
\usage{
conditional_DE_wrapper(time_object)
}
\arguments{
\item{time_object}{A timeseries object containing a DESeq2_obj}
}
\value{
The timeseries object with the conitional differential expression results
added to the DE_results slot of the object.
}
\description{
A wrapper function which performs conditional differential gene expression
analysis for every timepoint in the timeseries object

Conditional differential gene expression experiments involves the comparison
of the experiment vs the control. This is done at every timepoint.

The function calls the necessary sub functions for the differential expression
analysis based on the type needed (DESeq2 for RNAseq or limma for microarray)
}
