% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{convert_eb_res_to_DE_results}
\alias{convert_eb_res_to_DE_results}
\title{Format limma DE results}
\usage{
convert_eb_res_to_DE_results(
  time_object,
  eb_res,
  samples_used,
  exp_name,
  DE_type
)
}
\arguments{
\item{time_object}{A timeseries object}

\item{eb_res}{The empirical bayesian results}

\item{samples_used}{The samples used to obtain the eb_res object}

\item{exp_name}{The name which will be given to the results obtained by the function.}

\item{DE_type}{Either temporal or conditional for the type of differential gene
expression experiment performed}
}
\value{
The updated time object
}
\description{
Function which converts the empirical bayesian results into data frame format
for storage in a DE_results object.

The function also formats the data in regards to the column names in order
to ensure downstream compatibility.
}
