% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{create_conditional_heatmap_matrix}
\alias{create_conditional_heatmap_matrix}
\title{Create data for conditional heatmap}
\usage{
create_conditional_heatmap_matrix(time_object)
}
\arguments{
\item{time_object}{A time series object}
}
\value{
A list containing the main data matrix (count values) and the
three vectors described above.
}
\description{
This function is intended to be used within the \code{custom_heatmap_wrapper}

The function creates a list containing the main data matrix along with three vectors
l2fc_vector: A named vector containing the L2FC value for each gene of the matrix
group_vector: A vector containing the new names of the groups to represent the group
they are in
gene_vector: A named vector containing the gene names along with the group that
they are in.

The vectors are created to account for the possibility of duplicates. For example,
a gene may be significant in two timepoint comparisons, and therefore have two
different values. The vectors give the genes unique names based on their grouping
in order to enable them to have different values if necessary.
}
