% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{create_DE_data_results}
\alias{create_DE_data_results}
\title{Create and save DE results}
\usage{
create_DE_data_results(object, DE_type, exp_name, save_location)
}
\arguments{
\item{object}{A timeseries object}

\item{DE_type}{Character for the type of differential gene expression containing
the experiment of interest (conditional or temporal)}

\item{exp_name}{Character stating the name of the differential gene expression
of interest}

\item{save_location}{Character the location where the csv files will be saved.}
}
\value{
None
}
\description{
Function which saves the raw and significant version of the differential
gene expression data for a specific differential gene expression experiment
within a timeseries object
}
\examples{
TS_object<-create_example_object_for_R()
TS_object <- normalize_timeSeries_with_deseq2(time_object=TS_object)
#Perform conditional differential gene expression analysis
TS_object<-conditional_DE_wrapper(TS_object)
#Below function will save results to main directory
create_DE_data_results(TS_object,DE_type='conditional',exp_name='IgM_vs_LPS_TP_1',save_location='')

}
