% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{create_example_data_for_R}
\alias{create_example_data_for_R}
\title{Create a subset of count data for R examples (documentation)}
\usage{
create_example_data_for_R()
}
\value{
The smaller count matrix and sample data as a list
}
\description{
This function creates a subset of the AID count data in order to
present quick examples within the documentation
}
\examples{
example_dta<-create_example_data_for_R()
}
