% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{create_example_object_for_R}
\alias{create_example_object_for_R}
\title{Create a TimeSeries_Object from example data}
\usage{
create_example_object_for_R()
}
\value{
The example TimeSeries_Object
}
\description{
This function creates a TimeSeries_Object from example data
to use within documentation examples
}
\examples{
TS_object<-create_example_object_for_R()
}
