% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{create_raw_count_matrix}
\alias{create_raw_count_matrix}
\title{Prep TS matrix}
\usage{
create_raw_count_matrix(
  time_object,
  samp_data,
  path_to_data = NULL,
  limma_id_replace = "GeneName"
)
}
\arguments{
\item{time_object}{A timeseries object}

\item{samp_data}{dataframe containing the sample data}

\item{path_to_data}{The path towards the count files.}

\item{limma_id_replace}{The name of gene id's to use from the 'genes' dataframe
in the Elist}
}
\value{
The time_object, count matrix, and the slot to save it in
}
\description{
Function which prepares and ads the raw count matrix to the timeseries object

The function also differs in the preparation method depending onthe type of
differential gene expression method selected (DESeq2 or limma) as this affects
the input data. Where DESeq2 (RNAseq data) is stored as individual count file while
for limma the function expects a csv file with all samples and genes.
}
