% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{create_tables_genes_of_interest_DE}
\alias{create_tables_genes_of_interest_DE}
\title{Creates tables and plots for genes of interest}
\usage{
create_tables_genes_of_interest_DE(
  object,
  genes_of_interest,
  save_location,
  DE_type = c("conditional", "temporal"),
  log_tp = FALSE
)
}
\arguments{
\item{object}{A TimeSEries object}

\item{genes_of_interest}{A vector of strings indicating the genes of interest}

\item{save_location}{The directory in which the results will be saved}

\item{DE_type}{The differential gene expression axes for which the results will be retrieved}

\item{log_tp}{Boolean indicating if timepoints should be log10 transformed for the plot}
}
\value{
None
}
\description{
Function which retrieves, processes, and saves information regarding
genes which have been labelled as interesting by the user. The function will
retrieve the genes differential expression results on both the conditional and temporal
axes, then save these results in csv format. One csv file per gene of interest.

The function will also plot and save (as png) the trajectory of each gene
}
