% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortTimeSeriesData.R
\name{createCohortTimeSeriesArgs}
\alias{createCohortTimeSeriesArgs}
\title{Create a parameter object for the function getCohortTimeSeriesData}
\usage{
createCohortTimeSeriesArgs(cohortIds = -1, timeInterval = "month")
}
\arguments{
\item{cohortIds}{A vector of cohort Ids; set to -1 to use all cohorts in the cohort table}

\item{timeInterval}{The time interval (year, month, day) to use when extracting the cohort time series.}
}
\description{
Create a parameter object for the function getCohortTimeSeriesData
}
\details{
Create an object defining the parameter values for use with the
}
\seealso{
[getCohortTimeSeriesData]
}
