% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ocp.R
\name{createOcpModel}
\alias{createOcpModel}
\title{Create a Bayesian Online Changepoint model}
\usage{
createOcpModel(tsData, ocpArgs)
}
\arguments{
\item{tsData}{The time series data to use for fitting the model}

\item{ocpArgs}{The OCP linear model arguments. See \code{createSegmentedArgs}.
for more details.}
}
\description{
Create a Bayesian Online Changepoint model
}
\details{
Create a Bayesian Online Changepoint model using the time-series data
and arguments from \code{createSegmentedArgs}.
}
