% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmented.R
\name{createSegmentedModel}
\alias{createSegmentedModel}
\title{Create a segmented linear model}
\usage{
createSegmentedModel(tsData, segmentedArgs)
}
\arguments{
\item{tsData}{The time series data to use for fitting the model}

\item{segmentedArgs}{The segmented linear model arguments. See \code{createSegmentedArgs}.
for more details.}
}
\description{
Create a segmented linear model
}
\details{
Create a segmented linear model using the time-series data
and arguments from \code{createSegmentedArgs}.
}
