% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{createTsAnalysis}
\alias{createTsAnalysis}
\title{Create a TimeSeries analysis specification}
\usage{
createTsAnalysis(analysisId = 1, description = "", tsArgs)
}
\arguments{
\item{analysisId}{An integer that will be used later to refer to this specific
set of analysis choices.}

\item{description}{A short description of the analysis.}

\item{tsArgs}{An object representing the time series arguments to be used 
when calling the various time series model functions}
}
\description{
Create a TimeSeries analysis specification
}
\details{
Create a set of analysis choices, to be used with the [runTsAnalyses()] function.

Providing a NULL value for any of the argument applies the corresponding step will not be executed.
}
