% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{DE_plots_vignettes}
\alias{DE_plots_vignettes}
\title{Creates differential expression results for Vignettes}
\usage{
DE_plots_vignettes(object)
}
\arguments{
\item{object}{A timeseries object}
}
\value{
list of heatmaps (contains conditional and temporal plot)
}
\description{
The function will generate two summary heatmaps of the differential
gene expression results. The function was originally written to be used in Vignettes.
Full sets of differential gene expression results can be obtained with the
\code{plot_wrapper_DE_results} function
}
\examples{
TS_object<-create_example_object_for_R()
TS_object <- normalize_timeSeries_with_deseq2(time_object=TS_object)
#Perform conditional differential gene expression analysis
TS_object<-conditional_DE_wrapper(TS_object,vignette_run=TRUE)
TS_object<-temporal_DE_wrapper(TS_object,vignette_run=TRUE)
plot_list<-DE_plots_vignettes(TS_object)
}
