% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{exp_matrix}
\alias{exp_matrix}
\title{Fetches a data matrix from summarized experiments}
\usage{
exp_matrix(time_object, matrix_name)
}
\arguments{
\item{time_object}{A timeseries object}

\item{matrix_name}{Character for the name of the matrix. 'raw' or 'norm' are the standard
names found unless user modifications have been made}
}
\value{
The extracted matrix
}
\description{
Function which extracts a matrix from summarizedexperiments contained
within a TimeSeries_object. The matrix is extracted by name.
}
\examples{

TS_object <- new('TimeSeries_Object',
                 group_names=c('IgM','LPS'),group_colors=c("#e31a1c","#1f78b4"),DE_method='DESeq2',
                 DE_p_filter='padj',DE_p_thresh=0.05,DE_l2fc_thresh=1,
                 PART_l2fc_thresh=4,sem_sim_org='org.Hs.eg.db',Gpro_org='hsapiens')
TS_object <- TS_load_example_data(TS_object,'PBMC')
exp_matrix(TS_object,'raw')

}
