% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{exp_sample_data}
\alias{exp_sample_data}
\title{Fetches the sample data}
\usage{
exp_sample_data(time_object)
}
\arguments{
\item{time_object}{A timeseries object}
}
\value{
The data.frame format of sample data
}
\description{
Function which extracts the sample data from the SummarizedExperiment
contained within the TimeSerie_Object. The sample data is extracted and returned in
a data frame format
}
\examples{

TS_object <- new('TimeSeries_Object',
                 group_names=c('IgM','LPS'),group_colors=c("#e31a1c","#1f78b4"),DE_method='DESeq2',
                 DE_p_filter='padj',DE_p_thresh=0.05,DE_l2fc_thresh=1,
                 PART_l2fc_thresh=4,sem_sim_org='org.Hs.eg.db',Gpro_org='hsapiens')
TS_object <- TS_load_example_data(TS_object,'PBMC')
exp_sample_data(TS_object)

}
