% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_GO_results_functions.R
\name{find_merged_color}
\alias{find_merged_color}
\title{associate color to group for plotting}
\usage{
find_merged_color(group)
}
\arguments{
\item{group}{A vector of groups in the data}
}
\value{
new_col A vector of colors in the same order as the inputted group vector
}
\description{
Function which determines which color best represents color groupings (ex: blue/yellow)

The function takes pairs of two colors, finds the 6th in between color using
coloRampPalette, and continues doing so until all colors in the group have
been included
}
