% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_functions.R
\name{find_most_variable_cluster}
\alias{find_most_variable_cluster}
\title{Identify cluster with the most variability}
\usage{
find_most_variable_cluster(time_object, mean_ts_data)
}
\arguments{
\item{time_object}{A timeseries object}

\item{mean_ts_data}{A dataframe containing the mean trajectory for each cluster}
}
\value{
The name of the most variable cluster
}
\description{
Identifies the most variable cluster based on the mean trajectory
of each cluster. This cluster is identified for the purpose of illustration
in an rmarkdown file.
}
