% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_data}
\alias{get_data}
\title{Load the default time series data}
\usage{
get_data()
}
\value{
A time series object
}
\description{
Load the default time series data
}
\details{
This function allows the user to load a clean time series for
tests or training.
}
\examples{
get_data()
}
