% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortTimeSeriesData.R
\name{getCohortTimeSeriesData}
\alias{getCohortTimeSeriesData}
\title{Retrieve time series data for one or more cohorts}
\usage{
getCohortTimeSeriesData(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = NULL,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  outputFolder,
  databaseId,
  cohortTimeSeriesArgs
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema with write
privileges where temp tables can be created.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{The name of the cohort table.}

\item{outputFolder}{The location where the cohort time series data results will be written.}

\item{databaseId}{The database identifier for the time series data}

\item{cohortTimeSeriesArgs}{The cohort time series arguments. @seealso[createCohortTimeSeriesArgs] for more information.}
}
\description{
Retrieve time series data for one or more cohorts
}
\details{
This function will retrieve a data frame representing
entry events for one or more cohorts grouped into time intervals
as specified by the cohortTimeSeriesArgs
}
