% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{identify_problematic_combs}
\alias{identify_problematic_combs}
\title{Identify and remove bad combs}
\usage{
identify_problematic_combs(mat, min_shared_fields = 1)
}
\arguments{
\item{mat}{Matrix to investigate}

\item{min_shared_fields}{Minimum number of positions that are not NA in both
vectors in order not to flag the vector pair as problematic}
}
\description{
Hclust cannot handle matrices in which for some pairs of rows and columns,
only 1 or fewer shared values are non-NA. This function recurrently
identifies the most aggravating column/row, excludes that column/row and checks
whether more columns/rows need to be excluded


Function taken from github user–slagtermaarten
from the ComplexHeatmap issue number 155:
https://github.com/jokergoo/ComplexHeatmap/issues/155
}
