% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{launchShinyApp}
\alias{launchShinyApp}
\title{Launches the Shiny results viewer}
\usage{
launchShinyApp(outputFolder, shinySettings = list(dataFolder = outputFolder))
}
\arguments{
\item{outputFolder}{The root output folder that holds the individual database results}

\item{shinySettings}{This is a list() of settings passed to the Shiny application. At
the moment this is simply the output folder so likely no need to change this parameter.}
}
\description{
This function launches the Shiny web results viewer.
}
