% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{maplot_alt}
\alias{maplot_alt}
\title{Create an MA plot}
\usage{
maplot_alt(
  DE_res,
  genes_of_interest = c(),
  filter_choice,
  l2FC_thresh = 1,
  p_thresh = 0.05,
  plot_title = "MA plot"
)
}
\arguments{
\item{DE_res}{The differential expression results to be plotted}

\item{genes_of_interest}{A vector containing gene names to be labelled in the plot
To not label any genes, leave as default or provide an empty vector.}

\item{filter_choice}{Either padj or pvalue, the choice will be used to filter for
significance}

\item{l2FC_thresh}{The value used as the log2FoldChange threshold}

\item{p_thresh}{The value used to establish significance from adjusted pvalue}

\item{plot_title}{The title to be give to the plot}
}
\value{
None
}
\description{
The function creates a 'four quadrant' MA plot, where the FDR and log2FoldChange
thresholds dictate the significant up-regulated category, the significant downregulated
category, the significant low regulation category and the non-significant category

Genes of interest are labeled in a rectangle for visibility and will have the same
color as the category in which they are in.

The plot is created using ggplot2, to save a plot the ggsave() function is recommended
It is also recommended to use the following parameters to save the plot.
dpi=300 width=21 height=19 units='cm'
}
\examples{
TS_object<-create_example_object_for_R()
TS_object <- normalize_timeSeries_with_deseq2(time_object=TS_object)
#Perform conditional differential gene expression analysis
TS_object<-conditional_DE_wrapper(TS_object)
DE_res<-slot(TS_object,'DE_results')$conditional$IgM_vs_LPS_TP_1$DE_raw_data
ma_plot<-maplot_alt(DE_res = DE_res,filter_choice = 'padj')

}
