% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PART.R
\name{part_gprofiler_vignettes}
\alias{part_gprofiler_vignettes}
\title{Gprofiler analysis for Vignettes}
\usage{
part_gprofiler_vignettes(object, gene_vect)
}
\arguments{
\item{object}{A time series object}

\item{gene_vect}{The gene vector which will be submitted to gprofiler}
}
\value{
Either the gprofiler result, or the object updated with the example data
}
\description{
Function which runs the gprofiler function as a try catch. In the
event of a issue with the connection to host (unable to connect to gprofiler)
The function will fetch example data and use that instead.
}
