% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PBMC_pre_loaded}
\alias{PBMC_pre_loaded}
\title{Pre-computed PART, gprofiler,semantic data, and differential expression results for the PBMC dataset}
\format{
A list of lists
}
\source{
<https://www.github.com/Ylefol/TimeSeriesAnalysis>
}
\usage{
PBMC_pre_loaded
}
\description{
A list containing the gprofiler results (14 clusters) for the PBMC example.
}
\keyword{datasets}
