% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PBMC_TS_data}
\alias{PBMC_TS_data}
\title{RNAseq counts for AID activation cocktail in PBMC time series experiment}
\format{
A SummarizedExperiment containing the count matrix as the assay and the sample data as the colData
\describe{
  \item{sample_dta}{Dataframe indicating which count files belong to which group, timepoint, and replicate}
  \item{counts}{A matrix of all the count files}
}
}
\source{
<https://bioconductor.org/packages/release/bioc/html/SummarizedExperiment.html>

<https://www.github.com/Ylefol/TimeSeriesAnalysis>
}
\usage{
PBMC_TS_data
}
\description{
A data set with RNAseq results for various AID activation cocktails on a PBMC cell line.
Four groups exist: Control (no cocktail), LPS, IgM activation, and TGFb activation. Every group
has two replicates per timepoint. Timepoints are days 1, 3, and 9.
}
\keyword{datasets}
