% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{plot_custom_DE_heatmap}
\alias{plot_custom_DE_heatmap}
\title{Plots DE heatmaps}
\usage{
plot_custom_DE_heatmap(
  heat_mat,
  col_split,
  row_splits,
  l2fc_col,
  log_transform,
  legend_value = "counts",
  plot_file_name = "custom_heatmap",
  custom_width = 15,
  custom_height = 5
)
}
\arguments{
\item{heat_mat}{The main count matrix of the heatmap}

\item{col_split}{The named vector used to split the columns into it's segments}

\item{row_splits}{The named vector used to split the rows into the two groups}

\item{l2fc_col}{The log2foldchange vector used to create the histogram}

\item{log_transform}{boolean indicating if the results were log transformed or not}

\item{legend_value}{string indicating the type of value used
(counts for RNAseq, intensity value for microarray)}

\item{plot_file_name}{The name given to the saved heatmap}

\item{custom_width}{The width of the heatmap}

\item{custom_height}{The height of the heatmap}
}
\description{
The function is intended to be called within the \code{custom_heatmap_wrapper}

The function creates a segmented heatmap where rows are samples/patients and columns
are genes. The heatmap is column segmented based on the number of differential expression
analyses contained within it. A log2foldChange histogram is attached to the bottom
of the heatmap to indicate significance of the genes being visualized.

The column/segment legend is given at the bottom of the heatmap with the number
of genes in each group being also provided.

The gradient legend of the values within the heatmap is given on the right hand side.

The heatmap is saved in both png and svg format.
}
