% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{plot_gene_traj_single_tp}
\alias{plot_gene_traj_single_tp}
\title{Calculate trajectory of a single gene with no timepoints}
\usage{
plot_gene_traj_single_tp(time_object, target_gene, return_data = FALSE)
}
\arguments{
\item{time_object}{A timeseries object}

\item{target_gene}{String indicating the gene}

\item{return_data}{Boolean indicating of both data and plot should be returned}
}
\value{
dataframe with the trajectory data for the requested gene
}
\description{
Function takes in a target gene and will show the genes value for
each sample of the matrix.
}
\examples{
TS_object<-create_example_object_for_R()
TS_object <- normalize_timeSeries_with_deseq2(time_object=TS_object)
#Perform conditional differential gene expression analysis
TS_object<-conditional_DE_wrapper(TS_object,vignette_run=TRUE)
aicda_traj_dta<-calculate_gene_traj_data(TS_object,'AICDA')

}
