% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_GO_results_functions.R
\name{plot_MDS}
\alias{plot_MDS}
\title{Create MDS plot}
\usage{
plot_MDS(the_data, ggplot_version = FALSE)
}
\arguments{
\item{the_data}{The dataframe as created by \code{calculate_and_format_MDS}}

\item{ggplot_version}{Boolean indicating if a ggplot (static) version of the plot
should be created instead of a plotly version. By default is FALSE.}
}
\value{
p The plotly plot to enable either visualization or saving of the plot
}
\description{
The function will plot the GOs for each module while preserving their module name
as their color. The function will plot either an interactive MDS using plotly or a
static MDS/scatter plot using ggplot depending on the boolean given to the 'ggplot_version'
parameter.
}
\examples{
TS_object<-create_example_object_for_R()
TS_object <- normalize_timeSeries_with_deseq2(time_object=TS_object)
#Perform conditional differential gene expression analysis
TS_object<-conditional_DE_wrapper(TS_object,vignette_run=TRUE)
TS_object<-temporal_DE_wrapper(TS_object,do_all_combinations=TRUE,vignette_run=TRUE)
#Extract genes for PART clustering based on defined log(2)foldChange threshold
signi_genes<-select_genes_with_l2fc(TS_object)

#Use all samples, but implement a custom order. In this case it is reversed
sample_data<-exp_sample_data(TS_object)
TS_groups<-slot(TS_object,'group_names')
samps_2<-sample_data$sample[sample_data$group==TS_groups[2]]
samps_1<-sample_data$sample[sample_data$group==TS_groups[1]]

#Create the matrix that will be used for PART clustering
TS_object<-prep_counts_for_PART(object=TS_object,target_genes=signi_genes,scale=TRUE,target_samples=c(samps_2,samps_1))
TS_object<-compute_PART(TS_object,part_bootstrap=10,part_min_clust=10,dist_param="euclidean", hclust_param="average",vignette_run=TRUE)
TS_object<-run_gprofiler_PART_clusters(TS_object,vignette_run=TRUE) #Run the gprofiler analysis
#Results saved to created directory
gpro_res<-gprofiler_cluster_analysis(TS_object,'GO:BP',save_path=NULL)
GO_clusters<-gpro_res[['GO_df']]
sem_dta<-slot(TS_object,'sem_list')
plot_data<-calculate_and_format_MDS(GO_clusters,sem_dta)
plot_data<-merge_duplicate_modules(plot_data)
my_plot<-plot_MDS(plot_data)

}
