% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{plot_PCA_TS}
\alias{plot_PCA_TS}
\title{Plot timeseries PCA}
\usage{
plot_PCA_TS(time_object, exp_name = NULL, DE_type = NULL)
}
\arguments{
\item{time_object}{A timeseries object}

\item{exp_name}{The name of the experiment for which the PCA is plotted}

\item{DE_type}{Either conditional or temporal for the type of differential experiment
being plotted}
}
\value{
the pca_plot
}
\description{
For a DESeq2 analysis, the function which retrieves the variance stabililizing
transformation via the \code{vst()} of deseq2.
For a limma analysis, the function will calculate the PCs using plotMDS

If no DE_type is 'all', the function will plot the PCA for the normalized
data.

It then plots a labelled PCA plot in png format
}
