% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{plot_single_gene_traj}
\alias{plot_single_gene_traj}
\title{Plot gene trajectory}
\usage{
plot_single_gene_traj(mean_data, color_vector = NULL)
}
\arguments{
\item{mean_data}{dataframe containing mean reads per timepoint}

\item{color_vector}{a vector with color IDs where the names match the groups names
(by default it is Activated and Control)}
}
\value{
the ggplot2 object for the plot
}
\description{
The function plots the gene's trajectory across all timepoints for all the
groups contained in the dataframe. It performs a facet split based on the 'label'
column of the dataframe
}
\examples{
TS_object<-create_example_object_for_R()
TS_object <- normalize_timeSeries_with_deseq2(time_object=TS_object)
#Perform conditional differential gene expression analysis
TS_object<-conditional_DE_wrapper(TS_object)
aicda_traj_dta<-calculate_gene_traj_data(TS_object,'AICDA')
aicda_plot<-plot_single_gene_traj(aicda_traj_dta,TS_object@group_colors)

}
