% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{plot_wrapper_DE_results}
\alias{plot_wrapper_DE_results}
\title{Creates differential expression results}
\usage{
plot_wrapper_DE_results(
  object,
  DE_type,
  genes_of_interest = c(),
  results_folder = "TS_results/",
  adjust_missing_temp_samples = TRUE
)
}
\arguments{
\item{object}{A timeseries object}

\item{DE_type}{The type of differential gene expression done that the function
will use to create results. Either conditional or temporal}

\item{genes_of_interest}{A vector of genes of interest, can be empty (c())}

\item{results_folder}{The main folder where the results will be stored}

\item{adjust_missing_temp_samples}{Boolean indicating if missing temporal samples should be
compensated for (NA introduced). Currently, there is no reason to have this set to
False.}
}
\value{
None
}
\description{
The function will create a folder called DE_results_(conditonal or temporal),
it will then create a summary heatmap of all the differential expression experiments
done in that type.

The function will then iterate over the differential gene expression experiments done in
either DE_type (conditional or temporal) and creates two csv files, one containing
the raw differential expression data and the other is a subset of the raw data
containing only the significant genes
The function also creates a volcano plot and MA plot for each experiment

The individual differential expression results are stored in folders which
carry the name of the experiment
}
\examples{
TS_object<-create_example_object_for_R()
TS_object <- normalize_timeSeries_with_deseq2(time_object=TS_object)
#Perform conditional differential gene expression analysis
TS_object<-conditional_DE_wrapper(TS_object)
plot_wrapper_DE_results(TS_object,DE_type='conditional',results_folder = '')

}
