% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotSegmented}
\alias{plotSegmented}
\title{Plots time series data with change points computed with a segmented generalized
linear model}
\usage{
plotSegmented(
  tsData,
  model,
  plotTitle = "Segmented Regression",
  plotSubtitle = ""
)
}
\arguments{
\item{tsData}{The time series data used to train the model}

\item{model}{The trained segmented model}

\item{plotTitle}{The title for the plot}

\item{plotSubtitle}{The subtitle for the plot}
}
\description{
Plots time series data with change points computed with a segmented generalized
linear model
}
\details{
Create a plot to display the time series data along with any change points
detected by the segmented package
}
