% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Plots time series data}
\usage{
plotTimeSeries(
  tsData,
  applyFormatting = FALSE,
  plotTitle = "Time series",
  plotSubtitle = ""
)
}
\arguments{
\item{tsData}{The time series data used to train the model}

\item{applyFormatting}{When TRUE, a default formatting will be applied. Default is FALSE}

\item{plotTitle}{The title for the plot}

\item{plotSubtitle}{The subtitle for the plot}
}
\description{
Plots time series data
}
\details{
Create a plot to display the time series data and optionally formats the plot.
This function is used by the other plotting functions in this package
which will apply the formatting itself.
}
