% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PART.R
\name{prep_counts_for_PART}
\alias{prep_counts_for_PART}
\title{Format data for PART}
\usage{
prep_counts_for_PART(object, target_genes, scale, target_samples)
}
\arguments{
\item{object}{A time series object}

\item{target_genes}{Vector of genes to keep/use}

\item{scale}{Boolean indicating if the data should be scaled}

\item{target_samples}{Vector of samples to keep/use}
}
\value{
The timeseries object with the added PART count matrix
}
\description{
The function retrieves the counts from the time series object and filters
it to only contain relevant genes and samples. The data can also be scaled
if the parameter 'scale' is set to T
}
