% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{prep_limma_matrix}
\alias{prep_limma_matrix}
\title{Prepare data matrix from a Elist object}
\usage{
prep_limma_matrix(Elist_obj, replace_rows_with = NULL)
}
\arguments{
\item{Elist_obj}{A Elist object}

\item{replace_rows_with}{Character indicating with column (if any) will be used
to replace the rownames of the matrix in the Elist}
}
\value{
count_matrix
}
\description{
The function can replace ronames of the matrix by one contained in the 'gene'
container of the Elist
}
