% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{prep_tp_matrix}
\alias{prep_tp_matrix}
\title{limma timepoint matrix}
\usage{
prep_tp_matrix(
  time_object,
  groups_in_ts,
  target_tp = NULL,
  do_temporal = FALSE
)
}
\arguments{
\item{time_object}{A time series object}

\item{groups_in_ts}{Vector of the two groups/conditionals}

\item{target_tp}{Vector containing the two timepoints to compare}

\item{do_temporal}{boolean indication if the model matrix should be temporal or conditional}
}
\value{
a list contianing the model matrix and the subsets used
}
\description{
Function which prepares a timepoint model matrix for limma datasets

The function creates either a conditional matrix (do_temporal==FALSE) where
the model matrix will compare both conditions ex: WT vs Knockout
Or it will create a model for a temporal comparison, where it will compare the
two timpoints contained in 'groups_in_ts'

The function is meant to be used within a wrapper function
}
