% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{process_microarr_dta_limma}
\alias{process_microarr_dta_limma}
\title{Read and process microarray data using the limma package}
\usage{
process_microarr_dta_limma(
  micro_arr_path,
  micro_arr_source = "agilent",
  green.only = TRUE,
  backg_corr_meth = "normexp",
  back_corr_offset = 16,
  norm_arrays_meth = "quantile",
  ID_used = "GeneName"
)
}
\arguments{
\item{micro_arr_path}{String indicating the path to a folder containing all and only
microrna text}

\item{micro_arr_source}{The source of the microarray data. Choices can be seen using
limma documentation for the function \code{read.maimages}}

\item{green.only}{Boolean for the green only parameter of \code{read.maimages}}

\item{backg_corr_meth}{The background correction method with \code{backgroundCorrect}}

\item{back_corr_offset}{The offset integer to use with \code{backgroundCorrect}}

\item{norm_arrays_meth}{The normalization method to use with \code{normalizeBetweenArrays}}

\item{ID_used}{The gene IDs used for the Elist. The name given must match a column in the
'genes' dataframe of the data. If set to NULL, the first column is taken}
}
\value{
The processed and normalized Elist object
}
\description{
The function reads in the microarray data, removes the control probes, performs
background correction, normalization, and averages over irregular replicate probes (avereps)
}
