% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{runTsAnalyses}
\alias{runTsAnalyses}
\title{Run the time series analyses for the selected time series data set}
\usage{
runTsAnalyses(
  tsData,
  tsDataId,
  outputFolder = "./TimeSeriesAnalysis",
  tsAnalysisList
)
}
\arguments{
\item{tsData}{A vector representing the time series data}

\item{tsDataId}{An integer value to uniquely identify the time series data. Most often this value will be the cohort ID
that represents the cohort time series data.}

\item{outputFolder}{The folder where the time series models are saved}

\item{tsAnalysisList}{A list of time series analyses as specified by using the @seealso [createSegmentedArgs] and/or @seealso [createOcpArgs]}
}
\description{
Run the time series analyses for the selected time series data set
}
\details{
This function will run the list of time series analyses for the time series
data and save the output to the outputFolder.
}
