% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{select_genes_with_l2fc}
\alias{select_genes_with_l2fc}
\title{Select genes based on L2FC/FC}
\usage{
select_genes_with_l2fc(time_object, custom_l2fc_thresh = NULL)
}
\arguments{
\item{time_object}{A timeseries object}

\item{custom_l2fc_thresh}{A value indicating the log2FoldChange threshold, can be NULL}
}
\value{
A vector of significant genes whose absolute log2FoldChange is greater
than the inputed custom threhsold.
}
\description{
Function which goes through all differential gene expression experiments, both
conditional and temporal, and extracts all significant genes that have a
absolute log2FoldChange greater than the custom_l2fc_thresh
}
