% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{temporal_DE_wrapper}
\alias{temporal_DE_wrapper}
\title{Create temporal DE results}
\usage{
temporal_DE_wrapper(time_object)
}
\arguments{
\item{time_object}{A timeseries object containing a DESeq2_obj}
}
\value{
The timeseries object with the temporal differential expression results
added to the DE_results slot of the object.
}
\description{
Wrapper function which performs the temporal differential gene expression analyses

The temporal differential analyses is the comparison of each timepoint irrelevant
of their conition (experiment or control). Timepoints are compared with the next
immediate timepoint. For example, if there are three timepoints, two temporal
differential gene expressions will be performed: TP2 vs TP1 and TP3 vs TP2.

The function subsets the DESeq2_obj based on the necessary samples and adjusts
the condition of the object to reflect the comparison being done
}
