% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{TimeSeries_Object-class}
\alias{TimeSeries_Object-class}
\alias{TimeSeries_Object}
\title{The TimeSeries class}
\description{
The TimeSeries class is the main class for the timeseries pipeline
It is used to store the input data along with the final processed data and
any intermediate results that may be used for plot creation
}
\section{Slots}{

\describe{
\item{\code{exp_data}}{A summarized experiment which will contain the sample data and the count matrix}

\item{\code{group_names}}{A vector for the group names used, should be in order for
which the differential expression comparisons will occur, where the first
should be the experiment and second should be the control}

\item{\code{group_colors}}{A named vector indicating which colors should be associated to which group}

\item{\code{DE_method}}{Either limma or DESeq2 to indicate which differential expression method is used}

\item{\code{DE_p_filter}}{Character of paj or pvalue to use in the establishment of
significant differential expressed genes}

\item{\code{DE_p_thresh}}{A numeric value used along with the pvalue filter to establish
significance. Genes will be considered significant if they are below this threshold}

\item{\code{DE_l2fc_thresh}}{A numeric value for the log2FolChange threshold to establish
significance where the absolute value of the genes must be greater than the threshold}

\item{\code{Gpro_org}}{The organism used in gprofiler friendly format}

\item{\code{sem_list}}{The list containing the semantic similarity measures for each ontology}

\item{\code{DESeq2_obj}}{The normalized DESeq2 object}

\item{\code{limma_object}}{The EList limma object}

\item{\code{DE_results}}{A list of results for the different differential expression experiments
performed}

\item{\code{PART_l2fc_thresh}}{A integer indicating the log(2)foldchange threshold for genes
to be PART clustered}

\item{\code{PART_results}}{A list of results for the PART clustering analysis}

\item{\code{sem_sim_org}}{A string indicating the annotation DBI organism to use}

\item{\code{Gprofiler_results}}{A list of Gprofiler results for the PART clusters}
}}

\concept{objects}
