% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{TS_load_example_data}
\alias{TS_load_example_data}
\title{Add counts and sample data of examples to a TimeSeriesObject}
\usage{
TS_load_example_data(time_object)
}
\arguments{
\item{time_object}{A timeseries object}
}
\value{
The timeseries object with the raw count matrix added to it as well as the sample data
}
\description{
A function takes an existing TimeSeriesObject and adds the specified
example data to the object. Added data is the count matrix and sample data
}
\examples{


TS_object <- new('TimeSeries_Object',
                 group_names=c('IgM','LPS'),group_colors=c("#e31a1c","#1f78b4"),DE_method='DESeq2',
                 DE_p_filter='padj',DE_p_thresh=0.05,DE_l2fc_thresh=1,
                 PART_l2fc_thresh=4,sem_sim_org='org.Hs.eg.db',Gpro_org='hsapiens')
TS_object <- TS_load_example_data(TS_object)

}
