% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{write_example_data_to_dir}
\alias{write_example_data_to_dir}
\title{Prep example data in directory}
\usage{
write_example_data_to_dir(example_data, target_dir = NULL)
}
\arguments{
\item{example_data}{Either 'PBMC','MURINE', or 'CELEGANS' to select one of the three example
datasets available}

\item{target_dir}{Where the saves should be located, if NULL, it will be saved
to the main directory}
}
\value{
None
}
\description{
This function writes the example dataset to the 'data' directory
This presents the data in a way where users will understand the various files
they must provide for the TimeSeriesAnalysis pipeline as well as their format.

Having these files written as tab delimited and csv allows users to open them
and see the formatting expected by TimeSeriesAnalysis
}
\examples{
path <- tempfile()
bfc <- BiocFileCache(path, ask = FALSE)
write_example_data_to_dir('PBMC',target_dir=bfc@cache)


}
