% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_GO_results_functions.R
\name{calculate_and_format_MDS}
\alias{calculate_and_format_MDS}
\title{Calculate MDS matrix}
\usage{
calculate_and_format_MDS(GO_df, semantic_data, measure = "Wang")
}
\arguments{
\item{GO_df}{The dataframe of GOs as returned by \code{gprofiler_cluster_analysis}}

\item{measure}{the measuring method for semantic distance}

\item{sem_data}{semantic similarity data as created by the godata function}
}
\value{
res.mds A dataframe containing the MDS dimensions, along with groupings and
colors which are equivalent to the cluster colors
}
\description{
Function which calculates semantic distance and MDS from a dataframe of GOs

The function uses a GOSemSim object to calculates the termdistance, followed
by a MDS calculation.
The resulting dataframe is formatted to be used in \code{plot_MDS}

This function is used to create a GO term plot for WGCNA modules
}
