% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{calculate_cluster_traj_data}
\alias{calculate_cluster_traj_data}
\title{Calcualte trajectory of clusters}
\usage{
calculate_cluster_traj_data(object, custom_cmap = NULL, scale_feat = TRUE)
}
\arguments{
\item{object}{A timeseries object}

\item{scale_feat}{If the genes should be scaled/transformed using scale feature sum}
}
\value{
A dataframe containing the transformed or non-transformed gene values for
each cluster
}
\description{
Function which calculates the trajectory data for each cluster
Trajectory data is defined as the impact a gene has on time
This is obtained by performing a scale feature sum, where each
gene is 'equalized' by dividing the value of the gene (at each sample)
by the rowSum of the gene (the addition of the gene's values across all samples)
}
