% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{calculate_mean_cluster_traj}
\alias{calculate_mean_cluster_traj}
\title{Calculate mean trajectory of clusters}
\usage{
calculate_mean_cluster_traj(clust_traj_dta)
}
\arguments{
\item{clust_traj_dta}{gene trajectory data for each gene, obtained by \code{calculate_cluster_traj_data}}
}
\value{
A dataframe containing the mean cluster trajectories
}
\description{
Function which calculates the mean trajectory of each cluster
}
