% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PART.R
\name{compute_PART}
\alias{compute_PART}
\title{compute PART clusters}
\usage{
compute_PART(
  object,
  part_recursion = 100,
  part_min_clust = 10,
  dist_param = "euclidean",
  hclust_param = "average",
  custom_seed = NULL,
  custom_matrix = NULL,
  return_as_object = T
)
}
\arguments{
\item{object}{A timeseries object}

\item{part_recursion}{The number of recursions for PART calculation}

\item{part_min_clust}{The minimum number of genes per cluster}

\item{dist_param}{The distance parameter for clustering}

\item{hclust_param}{The hierarchical clustering method/parameter to be used}

\item{custom_seed}{An integer to be used to set a custom seed
which gives the option of reproducibility from the PART calculation}

\item{custom_matrix}{Allows the input of a custom matrix instead of taking it
from the object}

\item{return_as_object}{Boolean indicating if the results should be returned
within the submitted object or as a list}

\item{save_name}{the file name to save the PART data, if NULL no saving}
}
\value{
The timeseries object with the PART results added
}
\description{
The function set's a seed for reproducibility of results, it then uses the
part function from the clusterGenomics package to establish which genes belong
to what clusters. The clusters are then ordered using hierarchical clustering.
This may result in some PART cluster being split, therefore the order is adjusted
to put them together. We first order using hierarchical clustering for visual purposes.
}
