% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_GO_results_functions.R
\name{create_clustered_module_dataframe}
\alias{create_clustered_module_dataframe}
\title{Format clustered GO data for plotting}
\usage{
create_clustered_module_dataframe(cluster_df)
}
\arguments{
\item{cluster_df}{A dataframe containing the GO terms, their cluster, and module(s)
as created by \code{find_clusters_from_termdist}}
}
\value{
cluster_nb_df A dataframe containing GO terms for the clusters, their associated
cluster, the number of GO terms in each cluster as well as the module dispertion
of the GOs within the cluster
}
\description{
Function which counts the number of GOs in each cluster and determines how many
GOs of each module appears in each cluster
}
