% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_functions.R
\name{create_DEG_df}
\alias{create_DEG_df}
\title{Create summary table for DEGs}
\usage{
create_DEG_df(time_object)
}
\arguments{
\item{time_object}{A timeseries object}
}
\value{
The DEG table summarizing the results
}
\description{
Creates a table which summariezes the conditional and temporal
differetial gene expression results. The table shows the number of significant
DEGs found in each differential gene experiment performed. The table is created
for a Rmarkdown format report
}
