% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesData.R
\name{createCohortTimeSeriesArgs}
\alias{createCohortTimeSeriesArgs}
\title{Create a parameter object for the function getTimeSeriesData}
\usage{
createCohortTimeSeriesArgs(
  cohortIds = c(),
  timeUnits = "month",
  sortBy = "cohortStartDate"
)
}
\arguments{
\item{cohortIds}{A vector of cohort Ids; leave empty to use all cohorts in the cohort table}

\item{timeUnits}{The time units (year, month, day) to use when extracting the cohort time series.}

\item{sortBy}{The value from the cohort table to sort by: cohortStartDate or cohortEndDate}
}
\description{
Create a parameter object for the function getTimeSeriesData
}
\details{
Create an object defining the parameter values for use with the 
getTimeSeriesData
}
