% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_GO_results_functions.R
\name{custom_gpro_dotplot}
\alias{custom_gpro_dotplot}
\title{Create dotplot for gprofiler results}
\usage{
custom_gpro_dotplot(GO_clusters, ont, top_n)
}
\arguments{
\item{GO_clusters}{A dataframe with the group_name(cluster), term_name, -log10FDR,
and the term_size}

\item{ont}{The ontology used, it will be added to the title}

\item{top_n}{The number of top GOs used for the filter, will be added to the title}
}
\value{
Returns the ggplot2 object for the plot
}
\description{
Function which creates a dotplot for the provided gprofiler results

The function creates the ggplot2 object for a dotplot which illustrates the GOs
of the give clusters.
}
