% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{custom_heatmap_wrapper}
\alias{custom_heatmap_wrapper}
\title{Create heatmap for differential gene expression results}
\usage{
custom_heatmap_wrapper(
  time_object,
  DE_type,
  log_transform = TRUE,
  plot_file_name = "custom_DEG_heatmap",
  adjust_missing_temp_samples = TRUE
)
}
\arguments{
\item{time_object}{A time object}

\item{DE_type}{The type of heatmap being created (conditional or temporal)}

\item{log_transform}{Boolean indicating if the results should be log transformed
on the heatmap}

\item{plot_file_name}{The file name to be given to the heatmap once it is saved}

\item{adjust_missing_temp_samples}{Boolean indicating if missing temporal samples should be
compensated for (NA introduced). Currently, there is no reason to have this set to
False.}
}
\value{
none
}
\description{
A wrapper function which calls the necessary functions to create a heatmap
illustrating the significant DEGs of a given differential expression type (DE_type)
}
