% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGenomics.R
\name{doHclust}
\alias{doHclust}
\title{doHclust}
\usage{
doHclust(d, k, linkage)
}
\description{
Perform hierarchical clustering with given linkage, and apply horizontal cutting of tree into k clusters
}
