% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGenomics.R
\name{doKmeans}
\alias{doKmeans}
\title{doKmeans}
\usage{
doKmeans(X, k, nstart)
}
\arguments{
\item{X}{matrix of values}

\item{k}{value for kmeans}

\item{nstart}{Where to start}
}
\description{
Perform K-means clustering into k clusters
}
