% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_GO_results_functions.R
\name{dotplot_ancestors}
\alias{dotplot_ancestors}
\title{Create dotplot for ancestor query results}
\usage{
dotplot_ancestors(ancestor_df, use_names = F, enrichment_dta = T)
}
\arguments{
\item{ancestor_df}{The dataframe obtained from \code{find_relation_to_ancestors}}

\item{use_names}{Boolean indicating if the ancestor names or IDs should be used}

\item{enrichment_dta}{Boolean indicating if the plot has enrichment data}
}
\value{
The ggplot2 object for the created dotplot
}
\description{
Function which creates a dotplot of all the GO terms found to be related
to selected ancestors. The function expects the dataframe from \code{find_relation_to_ancestors}

The plot shows enirchment (up or down), to which ancestor a GO has been associated
as well as the size of the term. Each term is shown on the axis of the cluster
for wich it was found.
}
