% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGenomics.R
\name{findW}
\alias{findW}
\title{findW}
\usage{
findW(dX, K, cl.lab)
}
\arguments{
\item{K}{current cluster}

\item{cl.lab}{cluster labels}

\item{dx}{matrix for cluster}
}
\description{
Calulates the total within-cluster dispersion (W_K) for a k=1,...,K:
}
