% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGenomics.R
\name{gap}
\alias{gap}
\title{gap}
\usage{
gap(X, Kmax = 10, B = 100, ref.gen = "PC", cl.lab = NULL, ...)
}
\description{
gap finds the optimal number of clusters based on the "gap statistic" (Tibshirani et al. 2001)
}
