% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGenomics.R
\name{get.threshold}
\alias{get.threshold}
\title{get.threshold}
\usage{
get.threshold(X, q, ...)
}
\arguments{
\item{X}{value matrix}

\item{q}{q}
}
\value{
threshold
}
\description{
Help-functions only used by part:
Find a stopping threshold given the percentage of heights to be used in the dendrogram
}
