% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGenomics.R
\name{getDist}
\alias{getDist}
\title{getDist}
\usage{
getDist(X, dist.method, cor.method = "pearson")
}
\arguments{
\item{X}{Matrix used to calculate the distance}

\item{dist.method}{The distance method to use in the calculation}

\item{cor.method}{The correlation method to use. 'pearson' is the default,}
}
\value{
the found distance
}
\description{
Compute distance matrix for given distance measure; note: distance between rows!
}
\examples{
example_dta<-create_example_data_for_R()
X=as.matrix(example_dta$counts)
#Get distance matrix
dX <- getDist(X,dist.method='euclidean',cor.method='pearson')

}
