% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGenomics.R
\name{getPARTlabels}
\alias{getPARTlabels}
\title{getPARTlabels}
\usage{
getPARTlabels(clusters, minSize)
}
\arguments{
\item{clusters}{The found clusters}

\item{minSize}{The parameter for minimum cluster size}
}
\value{
PART labels
}
\description{
Retrives the labels for the clusters found by PART
}
