% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGenomics.R
\name{getReferenceW}
\alias{getReferenceW}
\title{getReferenceW}
\usage{
getReferenceW(X, Kmax, B, ref.gen, ...)
}
\arguments{
\item{X}{matrix of values}

\item{Kmax}{maximum number of clusters}

\item{B}{Number of recursions}
}
\value{
calculated Wb
}
\description{
Generate reference data sets and find Wb:
}
