% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_GO_results_functions.R
\name{gprofiler_cluster_analysis}
\alias{gprofiler_cluster_analysis}
\title{Plot gprofiler results}
\usage{
gprofiler_cluster_analysis(
  object,
  ontology,
  save_path = NULL,
  return_specific_cluster = NULL
)
}
\arguments{
\item{object}{A time series object}

\item{ontology}{the ontology that will be returned in the dataframe (ex: GO:BP)}

\item{save_path}{The folder path to save results if gprofiler results are to be saved}

\item{return_specific_cluster}{String to return the gost plot of a specific cluster}
}
\value{
GO_df A dataframe containing all the GOs (ID) found, their cluster, and the term name
}
\description{
Creates a dataframe of GO's based on the gprofiler results stored in a time series
object. The dataframe is specific to the requested ontology

If save_path is not null, the gprofiler results will be saved to the designated
location in both csv and html (interactive plot) format.
}
