% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{log_transform_l2fc_vect}
\alias{log_transform_l2fc_vect}
\title{Log transform L2FC/FC values}
\usage{
log_transform_l2fc_vect(l2fc_vector)
}
\arguments{
\item{l2fc_vector}{The log2FoldChange vector to be log transformed}
}
\value{
The log transformed vector
}
\description{
The function first identifies the location of negative values. It then performs
the log transformation on the absolute values of the vector.
Lastly it returns originally negative values to negative values
This is done as log transformation does not work on negative values, therefore a
work around had to be done.

This function is intended to be used within \code{custom_heatmap_wrapper}
}
