% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_GO_results_functions.R
\name{merge_duplicate_modules}
\alias{merge_duplicate_modules}
\title{merge_MDS modules}
\usage{
merge_duplicate_modules(the_data)
}
\arguments{
\item{the_data}{dataframe as returned by \code{calculate_and_format_MDS}}
}
\value{
the_data The updated inputted dataframe
}
\description{
Function finds and merges GOs from separate modules

If a go appears as a part of two or more modules, the function will remove the duplicate
GOs and keep a single version which has all the relevant modules as it's source
ex: blue/yellow/brown if the GO was found in the blue, yellow, and brown cluster
It also finds the color which best mixes the colors of the cluster
}
