% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGenomics.R
\name{part}
\alias{part}
\title{part}
\usage{
part(X, Kmax = 10, minSize = 8, minDist = NULL, cl.lab = NULL, ...)
}
\arguments{
\item{X}{matrix of values}

\item{minSize}{minimum cluster size}

\item{minDist}{Minimum distance between clusters}

\item{cl.lab}{cluster labels}

\item{kmax}{maximum number of clusters}
}
\value{
list of identified clusters
}
\description{
Main function for PART:
}
