% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PART.R
\name{part_gprofiler_vignettes}
\alias{part_gprofiler_vignettes}
\title{Gprofiler analysis for Vignettes}
\usage{
part_gprofiler_vignettes(object)
}
\arguments{
\item{object}{A time series object}
}
\value{
Updated object with gprofiler results
}
\description{
Function which loads gprofiler results for vignettes/examples
}
