% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGenomics.R
\name{PartRec}
\alias{PartRec}
\title{PartRec}
\usage{
PartRec(X, Kmax, ind, cl.lab = NULL, ...)
}
\arguments{
\item{X}{Matrix of values}

\item{Kmax}{max number of cluster}

\item{ind}{number of clusters}

\item{cl.lab}{cluster label}
}
\description{
The recursive function used by PART clustering
}
