% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PBMC_pre_loaded}
\alias{PBMC_pre_loaded}
\title{Pre-computed PART and gprofiler results for the PBMC dataset}
\format{
A list of lists
}
\source{
<https://www.github.com/Ylefol/TimeSeriesAnalysis>
}
\usage{
PBMC_pre_loaded
}
\description{
A list containing the gprofiler results (14 clusters) for the PBMC example.
This data exists to allow for the creation of Vignettes in the event of a network failure.
Also contains a list of lists with various PART results, from the inputed matrix to the
computed clusters and the various parameters used for the clustering.
}
\keyword{datasets}
