% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_GO_results_functions.R
\name{plot_ancestor_clust_MDS}
\alias{plot_ancestor_clust_MDS}
\title{Plot nearest ancestor MDS (clustered GOs)}
\usage{
plot_ancestor_clust_MDS(the_data, use_name = F)
}
\arguments{
\item{the_data}{The dataframe as created by \code{calculate_and_format_MDS}}
}
\value{
p The plotly plot to enable either visualization or saving of the plot
}
\description{
Function which plots the MDS for clusters/nearest ancestor of the ancestor curation
dataframe
}
