% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{plot_cluster_traj}
\alias{plot_cluster_traj}
\title{Plot cluster trajectories}
\usage{
plot_cluster_traj(
  object,
  ts_data,
  ts_mean_data,
  num_col = 4,
  rem_legend_axis = F,
  log_TP = F,
  title_text_size = 14
)
}
\arguments{
\item{object}{A timeseries object}

\item{ts_data}{The trajectory data for all clusters being calculated
The data is calculated/obtained from \code{calculate_cluster_traj_data} function}

\item{ts_mean_data}{The trajectory data for all clusters being calculated
The data is calculated/obtained from \code{calculate_cluster_traj_data} function}

\item{num_col}{Integer stating the number of columns for the plots.}

\item{rem_legend_axis}{Boolean indicating if the legend and axis titles should be removed}

\item{log_TP}{Boolean indicating if timepoints should be log transformed}

\item{title_text_size}{Integer indicating what the font size of the titles should be in the facets}
}
\value{
A ggplot2 object for the cluster trajectory plot performed
}
\description{
Function which creates a grid of cluster trajectories. Each cluster is split
into two subplots, one for the control and the other for the experiment.

Individual gene trajectories are plotted for each cluster along with a large
gray line for the mean cluster trajectories.
}
