% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_GO_results_functions.R
\name{plot_MDS}
\alias{plot_MDS}
\title{Create MDS plot}
\usage{
plot_MDS(the_data)
}
\arguments{
\item{the_data}{The dataframe as created by \code{calculate_and_format_MDS}}
}
\value{
p The plotly plot to enable either visualization or saving of the plot
}
\description{
The function will plot the GOs for each module while preserving their module name
as their color.
}
