% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{prep_sample_data}
\alias{prep_sample_data}
\title{Sample data set-up}
\usage{
prep_sample_data(path, group_names)
}
\arguments{
\item{path}{The path that will be read to obtain the sample data}

\item{group_names}{The group names selected by the user}
}
\value{
the subsetted sample file based on the selected groups
}
\description{
A simple function to read and subset the sample_data created by the user.
}
\examples{
path <- tempfile()
bfc <- BiocFileCache(path, ask = FALSE)
write_example_data_to_dir('PBMC',target_dir=bfc@cache)
my_path_data<-paste0(bfc@cache,'/data/PBMC/raw_counts_TS')
my_path_sample_dta<-paste0(bfc@cache,'/data/PBMC/sample_file.csv')
prep_sample_data(my_path_sample_dta,c('IgM','LPS'))

}
