% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{prepare_heat_data}
\alias{prepare_heat_data}
\title{Process data for DE heatmap}
\usage{
prepare_heat_data(matrix_list, log_transform)
}
\arguments{
\item{matrix_list}{The heat_list produced by \code{create_heatmap_matrix}}

\item{log_transform}{If a log transformation of the data is necessary}
}
\value{
An updated version of the inputted matrix_list
}
\description{
Function which adjusts and renames certain elements of the heat_list
produced by \code{create_heatmap_matrix}

It performs a log_transformation if necessary and converts the vectors into
factors to be used for annotation purposes within the heatmap

This function is intended to be used within the \code{cusom_heatmap_wrapper}
}
\examples{
TS_object<-create_example_object_for_R()
TS_object <- normalize_timeSeries_with_deseq2(time_object=TS_object)
#Perform conditional differential gene expression analysis
TS_object<-conditional_DE_wrapper(TS_object,vignette_run=TRUE)
heat_dta<-create_conditional_heatmap_matrix(TS_object)
heat_dta<-prepare_heat_data(heat_dta,log_transform=TRUE)

}
