% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PART_results_functions.R
\name{prepare_top_annotation_PART_heat}
\alias{prepare_top_annotation_PART_heat}
\title{Create top annotation for PART heatmap}
\usage{
prepare_top_annotation_PART_heat(object)
}
\arguments{
\item{object}{A time series object}
}
\value{
A list containing the two types of top annotaiton and the named vector used to create the column splits
}
\description{
This function is intended to be used within \code{PART_heat_map}

The top annotation shows annotation blocks for the groups as well as annotation
gradients for the time points.
Two sets of top annotations are created, one labelled and one unlabeled.
In addition, the column splits are created based on the replicates
}
