% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_GO_results_functions.R
\name{read_gprofiler_results}
\alias{read_gprofiler_results}
\title{Read gprofiler data from results folder}
\usage{
read_gprofiler_results(
  res_location = "TS_results/",
  ont = "REAC",
  top_n = NULL
)
}
\arguments{
\item{res_location}{The location of the gprofiler_results folder}

\item{ont}{The ontology to be filtered for}

\item{top_n}{The number of top GOs to be filtered for}
}
\value{
A dataframe with the requested information
}
\description{
Function which reads the grprofiler data results from the time series analysis

The function reads the information for each cluster and filters for the required
columns, ontology, and number of GOs. It then adds a -log10 FDR
}
