% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PART.R
\name{run_gprofiler_PART_clusters}
\alias{run_gprofiler_PART_clusters}
\title{Gprofiler analysis}
\usage{
run_gprofiler_PART_clusters(object)
}
\arguments{
\item{object}{A time series object}
}
\value{
The updated object with the Gprofiler results
}
\description{
Function which runs Gprofiler for PART clusters and stores the results
in the appropriate slot of the time object
}
