% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_GO_results_functions.R
\name{SS_GO_clusters}
\alias{SS_GO_clusters}
\title{Find semantic similarity between clustered GOs}
\usage{
SS_GO_clusters(semantic_data, the_clusters, selected_ont, distance, measure)
}
\arguments{
\item{semantic_data}{semantic similarity data as created by the godata function}

\item{the_clusters}{GO terms with associated clutsers as created by \code{find_clusters_from_termdist}}

\item{selected_ont}{The ontology to be used}

\item{distance}{The distance algorithm to use (ex:'BMA')}

\item{measure}{The measuring algorithm to use (ex: 'Wang')}
}
\value{
values A matrix of semantic distance for the clustered GOs
}
\description{
Function which clusters GO terms and returns the result in dataframe format

This function uses the Wang measurement method with the ward.D2 aggregation method
clustering parameters are pre-set within the function.

Function was adapted from the calculate_SS function from the visEAGO package
}
