% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_GO_results_functions.R
\name{wrapper_ancestor_curation_plots}
\alias{wrapper_ancestor_curation_plots}
\title{Create ancestor query plots}
\usage{
wrapper_ancestor_curation_plots(
  GO_df,
  sem_data,
  use_names = TRUE,
  target_dir = "TS_results/",
  return_plot = FALSE
)
}
\arguments{
\item{GO_df}{The dataframe of GOs as returned by \code{find_relation_to_ancestors}}

\item{sem_data}{semantic similarity data as created by the godata function}

\item{use_names}{boolean indicating if names or IDs should be used}

\item{target_dir}{string indicating the save location of the plots}

\item{return_plot}{boolean indicating if the plot should be returned}
}
\value{
if specified, will return a list containing the ggplot2 object for the dotplot
and the plotly object for the MDS plot.
}
\description{
Wrapper function which makes the necessary calls to plots ancestor curation plots

These plots find all terms which are children of the provided ancestors within the
specified ontology.

The function will plot and save a dotplot as well as an interactive MDS plot
}
