% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_GO_results_functions.R
\name{wrapper_MDS_and_MDS_clusters}
\alias{wrapper_MDS_and_MDS_clusters}
\title{Create MDS and nearest ancestor MDS plots}
\usage{
wrapper_MDS_and_MDS_clusters(
  GO_clusters,
  sem_data,
  sem_ontology,
  target_dir = "TS_results/gprofiler_results/",
  return_plot = FALSE
)
}
\arguments{
\item{GO_clusters}{The dataframe of GOs contained in the 'Gpro_org' slot of a time_object}

\item{sem_data}{semantic similarity data as created by the godata function}

\item{sem_ontology}{The ontology that was used to calculate the semantic data
and that will be plotted}

\item{target_dir}{string indicating the location where the results will be saved}

\item{return_plot}{boolean indicating if the plots should be returned}
}
\value{
if specified, will return a list containing the two plotly objects for
both the GO term MDS and clustered GO term MDS.
}
\description{
Wrapper function which makes the necessary calls to plot a multi dimensional
scaling (MDS) plot for all of the terms found in each cluster (for a specified ontology).
The function calculates and plots a nearest ancestor version of the MDS using the
BMA and WANG approach. Each term is brought up to it's nearest ancestor.

Each plot is saved as an interactive plot in TS_results
}
